requirements.txt
dotenv
langchain_community
pytest
langchain_deepseek
openrouter
